Description: optimize font memory IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I69CGJ Feature or Bugfix: feature Binary Source:No Signed-off-by: lancer <haoshuo@huawei.com>

+++ frameworks/common/typed_text.cpp
@@ -522,7 +522,7 @@ bool TypedText::IsColourWord(uint32_t codePoint, uint16_t fontId, uint8_t fontSi
         hasColor = true;
     } else {
 #if defined(ENABLE_MULTI_FONT) && ENABLE_MULTI_FONT
-        uint8_t* searchLists = nullptr;
+        uint16_t* searchLists = nullptr;
         int8_t listSize = UIMultiFontManager::GetInstance()->GetSearchFontList(fontId, &searchLists);
         if ((listSize > 0) && (searchLists != nullptr)) {
             int8_t currentIndex = 0;
